sync: avoid locking individual per-P shared pools if they are likely to be empty
The check for emptyness is racy because it's not done with the mutex locked. While this
is obviously undesirable, it's required to avoid the locking and it does not affect
the correctness of the program, because the sync.Pool docs make no promises about what
Get returns.

This is especially important when the Pool is operating mostly empty because getSlow is
O(N) in the number of procs.

The potential outcomes of the check-before-lock scheme are as follows:
- the check correctly predicts that the pool is empty: in this case we simply avoid
  a Lock/Unlock of the mutex
- the check correctly predicts that the pool is not empty: in this case we do as we
  were doing before this commit: Lock, pop from pool, Unlock
- the check wrongly predicts that the pool is not empty: in this case we also do as
  we were doing before this commit: Lock, skip popping, Unlock
- the check wrongly predicts that the pool is empty: in this case we skip the Lock/
  pop/Unlock, thereby choosing to not return an item from the per-P shared pool. This
  is suboptimal because it can lead to an increase in memory allocations but this
  should be a rare occurrence because there are multiple per-P shared pools, so
  the chances of this misprediction mattering should be small in practice

In addition to the main change above, also refactor some of the Pool methods to minimize
the number of branches and add a Pool benchmark covering the case in which more elements
are Get() from the pool than they are Put() (BenchmarkPoolUnderflow)

name             old time/op  new time/op  delta
Pool-4           14.5ns ± 3%  14.2ns ± 2%   -1.64%  (p=0.023 n=9+10)
PoolOverflow-4   1.99µs ±12%  1.78µs ± 1%  -10.62%  (p=0.000 n=10+8)
PoolUnderflow-4   152ns ± 6%    30ns ± 1%  -80.00%  (p=0.000 n=10+8)

Change-Id: I5b8e59a57c2747a8081585e8243bbcdc25b65406