sync: inline the fast path of Mutex.Lock(), Mutex.Unlock() and Once.Do()
Split the slow path in separate lockSlow/unlockSlow/doSlow and invert
the if condition. The latter change is so that after inlining the fast
path immediately follows the if-condition test and branch, and the slow
path is at the end of the function. This is beneficial because:
1) the fast path should be hotter, so keeping it close to the happy
path while pushing the slow path away should improve i-cache usage.
2) (less important) the static branch prediction on x86 is not taken
for forward jumps: with this change the slow path is after the forward
jump. This is likely not so relevant for recent processors, but can't
hurt.

Note: this fundamentally requires mid-stack inlining to be enabled
to be effective.

Change-Id: I796d94cfec396d5838792b6c5d1c832fcf1edce9