runtime: predict stack sizing
Goroutines can spend significant time in morestack/newstack if the dynamic
call tree is large, or if a specific call tree makes heavy use of the stack
to allocate significant amount of space for temporary variables.

This CL adds a simple stack size predictor based on the pc of the go statement
that starts each goroutine. This approach is predicated on the assumption that
a specific go statement in the program will mostly result in the resulting
goroutine mostly executing the same dynamic call tree (more precisely, dynamic
call trees with similar stack sizing requirements).

The way it works is by embedding in each P a small prediction table with 32 slots.
When a go statement is executed, the pc of the go statement is hashed with a per-P
seed to pick one of the slots. Each slot is a single byte containing a simple
running estimator. The result of the estimation is _StackMin << n, with n in the
range 0 to 15 (i.e. 2KB to 64MB), and it is used to allocate a stack of the
appropriate size. In newstack, called from morestack when we need to allocate a
new stack, we record the highest stack size (highwater) used by each goroutine
(this highwater is stored in the g struct, but thanks to existing padding this
additional 1-byte field does not cause the g struct to increase in size). When a
goroutine exits, the estimator for the pc of the statement that started that
goroutine is updated using the highwater value recorded by the exiting goroutine.

The current estimation scheme is not precise for multiple reasons.
First, multiple pcs could map to the same slot in the per-P table. This is not a
significant problem if we assume that the conflicting pcs are not executed with
the same frequency as in this case the estimator will still converge, albeit more
slowly, to the correct value. Furthermore, each P uses a different seed when
hashing the pc, and the seeds themselves are periodically reset (currently, as
a result of GC - although this is not the only available option).
Second, the highwater mechanism partially relies on stackguard0; the current
runtime sometimes mangles stackguard0 (e.g. when a goroutine needs to be pre-
empted) and this can lead to the highwater of a goroutine to be lower than it
should have been: this also should not lead to problems, apart from the estimator
taking longer to converge to the true value.

The stack size prediction mechanism is disabled by default, and can be enabled by
setting GOEXPERIMENT=predictstacksize.

This CL is currently in a PoC state. It passes all tests locally, and shows
significant promise in the included benchmark, where enabling stack size
prediction leads to a doubling of the performance for medium-large stack sizes.
It is still missing tests for the new feature pending discussion about the
proposed approach.

DO NOT SUBMIT

Fixes #18138

Change-Id: Id2f617e39bbd7ed969d35e1f231ab61c207fa572