runtime: provide a fast path for GOMAXPROCS(0)
Currently reading the value of GOMAXPROCS involves locking/unlocking
the global scheduler mutex. This made sense at the time of the C->Go
translation, but the runtime has evolved in the meantime and this is
seemingly not required anymore (many other places in the runtime
perform reads of gomaxprocs without locking).

This gives us the opportunity to provide an inlineable fast path,
that effectively turns a call to GOMAXPROCS(0) into a single inline
load from memory (e.g. `MOVSXD runtime.gomaxprocs(SB), <reg>` in
amd64).

Precise benchmark results are TBD, but speedups observed range from
10x to 350x faster, depending on the level of concurrency.

Having a fast way to read the current value of GOMAXPROCS is
important as it can be used to dynamically decide whether to use
a serial or parallel algorithm to process varying amounts of data,
e.g. when processing large slices.

Change-Id: I38be2ff0ec46f9c6813092ed13e000ca9c5c434c