runtime: update runtime.freegc to support heap objects containing pointers
WIP.

This CL is part of a series of CLs to that attempt to reduce how much
work the GC must do. See CL 700255 for an overall design document.

The earlier CL 673695 added runtime.freegc, but only for noscan
heap objects (objects that do not themselves contain pointers):

 func freegc(ptr unsafe.Pointer, uintptr size, noscan bool)

This CL updates freegc to handle scan heap objects (objects
containing pointers), along with corresponding updates elsewhere
in the runtime.

See the overall design in CL 700255 for more details.

Updates #74299

Change-Id: Ic285917ac83372ef20488c897132874ace8bea86